api/requirements.txt
Django==3.0.4
djangorestframework==3.11.0
markdown==3.2.1
django-filter==2.2.0
django-cors-headers==3.2.1
coreapi==2.3.3
psycopg2==2.8.5
PyMySQL==0.10.1
jsonfield==3.1.0
docxtpl==0.11.2
lxml==4.6.1
python-docx==0.8.10
html2docx==1.3.1
